[#13] Migrate to `zcash/librustzcash` revision with NU5 awareness (#20)
* Migrate to `zcash/librustzcash` revision with NU5 awareness

This enables mobile wallets to send transactions after NU5 activation.

* Commit `Cargo.lock` to pin Rust dependencies

This repository produces a binary output (a static FFI library that has
evaluated and internalised all of its dependency relationships), so it
should follow the convention for Rust binaries.

* Migrate to `zcash/librustzcash` revision with v5 tx parsing

This doesn't affect Sapling note detection (which uses Compact Blocks),
but enables the enhancement logic to decrypt and store the corresponding
transactions for displaying memo fields.

* Ignore `products/` and `.swiftpm`

* Store raw transaction data instead of reserializing.

Co-authored-by: Francisco Gindre <francisco.gindre@gmail.com>
Co-authored-by: Kris Nuttycombe <kris@nutty.land>