Use unified spending keys instead of Sapling extended spending keys.
This also updates `zcashlc_create_to_address` to take a unified spending
key for spending authority instead of a Sapling extended spending key.
Also, the unified spending key is now required for shielding.  This
removes the need to separately derive the transparent spending key, and
provides overall a more consistent API experience, such that the caller
always works in terms of a single, opaque, binary representation of the
spending key.

This also removes direct derivation of addresses from the seed, spending
key, and viewing key, in favor of using the wallet backend to manage
addresses. This is necessary to ensure that diversified addresses
produced by this API are properly recorded in the wallet database, so
that the transparent receivers associated with newly derived diversified
addresses can be correctly monitored for incoming funds.

After this change, the direct dependency on hdwallet is no longer needed.