Return a structured result from `zcashlc_tor_lwd_conn_check_single_use_taddr`
Returning a `bool` is insufficient because it does not provide an
adequate path for determining whether an error occurred, and it also
does not provide enough information for the wallet to then initiate a
shielding action for the funds discovered on an address. We now return
structured data such that a null pointer is returned in the case of an
error, and an enumerated type containing the address at which funds were
discovered otherwise.